Remove json-bigint dependency
