+++ lib/utils.js
@@ -298,6 +298,67 @@ function compareVersions(version1, version2, operator) {
   }
 }
 
+/**
+ * Parse JSON string with support for large integers (beyond Number.MAX_SAFE_INTEGER).
+ * Large integers are automatically converted to strings to preserve precision.
+ *
+ * For Node.js >= 21.0.0, uses native JSON.parse() with context parameter.
+ * For older versions, preprocesses the JSON string with regex to convert unsafe integers.
+ *
+ * Replaces: json-bigint package
+ *
+ * @param {string} str - JSON string to parse
+ * @returns {any} Parsed JSON object with unsafe integers as strings
+ *
+ * @example
+ * parseJSONWithBigInt('{"value": 9007199254740992}')
+ * // Returns: { value: '9007199254740992' }
+ *
+ * parseJSONWithBigInt('{"value": 100}')
+ * // Returns: { value: 100 }
+ */
+function parseJSONWithBigInt(str) {
+  const contextSupportedVersion = '21.0.0.0';
+  const currentVersion = process.version;
+  const isContextSupported = compareVersions(
+    currentVersion.substring(1, currentVersion.length),
+    contextSupportedVersion,
+    '>='
+  );
+
+  // For nodejs >= 21.0.0, we leverage context parameter to
+  // convert unsafe integer to string.
+  // See https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/parse
+  if (isContextSupported) {
+    return JSON.parse(str, (key, value, context) => {
+      if (
+        Number.isInteger(value) &&
+        !Number.isSafeInteger(Number(context.source))
+      ) {
+        return context.source;
+      }
+      return value;
+    });
+  }
+
+  // For older Node.js versions, preprocess the JSON string to convert
+  // large integers (outside the safe integer range) to strings before parsing.
+  // This replaces the functionality of json-bigint package.
+  const processed = str.replace(
+    /:\s*(-?\d+)(?=\s*[,}\]])/g,
+    (match, number) => {
+      const num = Number(number);
+      // If the number is not safe, keep it as a string
+      if (!Number.isSafeInteger(num)) {
+        return `: "${number}"`;
+      }
+      return match;
+    }
+  );
+
+  return JSON.parse(processed);
+}
+
 module.exports = {
   // General utilities
   detectUbuntuCodename,
@@ -320,5 +381,7 @@ module.exports = {
   removeSync,
   readJsonSync,
 
+  // Version and JSON utilities
   compareVersions,
+  parseJSONWithBigInt,
 };
