+++ package.json
@@ -77,7 +77,6 @@
     "@rclnodejs/ref-struct-di": "^1.1.1",
     "bindings": "^1.5.0",
     "debug": "^4.4.0",
-    "json-bigint": "^1.0.0",
     "node-addon-api": "^8.3.1",
     "walk": "^2.3.15"
   },
